update numpy deprecations

+++ optical_flow_estimation/raft/raft.py
@@ -265,8 +265,8 @@ def grid_sample(image, grid, padding_mode='border', align_corners=False):
 
     c = np.zeros_like(y) + np.arange(C)[np.newaxis, :, np.newaxis, np.newaxis]
     n = np.zeros_like(y) + np.arange(N)[:, np.newaxis, np.newaxis, np.newaxis]
-    c = c.astype(np.int)
-    n = n.astype(np.int)
+    c = c.astype(int)
+    n = n.astype(int)
 
     # Unnormalize with align_corners condition
     ix = grid_sampler_compute_source_index(x, W, align_corners)
@@ -287,10 +287,10 @@ def grid_sample(image, grid, padding_mode='border', align_corners=False):
     # Get values of the image by provided x0,y0,x1,y1 by channel
 
     # image, n, c, x, y, H, W
-    x0 = x0.astype(np.int)
-    y0 = y0.astype(np.int)
-    x1 = x1.astype(np.int)
-    y1 = y1.astype(np.int)
+    x0 = x0.astype(int)
+    y0 = y0.astype(int)
+    x1 = x1.astype(int)
+    y1 = y1.astype(int)
     Ia = safe_get(image, n, c, x0, y0, H, W, padding_mode)
     Ib = safe_get(image, n, c, x0, y1, H, W, padding_mode)
     Ic = safe_get(image, n, c, x1, y0, H, W, padding_mode)
